use process.stdout/stderr.write in place of console.log/error
We're mixing usage of console.log with process.stdout.write.
Worse, we sometimes console.log('...\n'), although console.log
appends a newline.  And sometimes it's intentional, because we want
two newlines, while other times it isn't.

So it's a bit of a mess, and to clean it up, this commit uses
process.stdout.write everywhere, explicitly apppending the desired
number of newlines (except when dumping errors, in which case
it uses process.stderr.write).

Along the way, I wrapped some long strings and added info/warning
prefix chars (i.e. ℹ, ⚠) to a few messages that deserved them.